Restructure inference module
